Add new abstract class which allow to easy write a Mailbox implementation and not depend on the MessageMapper (IMAP-202)
git-svn-id: https://svn.apache.org/repos/asf/james/imap/trunk@987619 13f79535-47bb-0310-9956-ffa450edef68

+++ store/src/main/java/org/apache/james/imap/store/ResultIterator.java
@@ -93,7 +93,7 @@ public class ResultIterator<Id> implements Iterator<MessageResult> {
 
         private final Date internalDate;
 
-        private final int size;
+        private final long size;
 
         private final long uid;
 
@@ -101,7 +101,7 @@ public class ResultIterator<Id> implements Iterator<MessageResult> {
                 final MailboxException exception) {
             super();
             internalDate = message.getInternalDate();
-            size = (int)message.getDocument().getFullContentOctets();
+            size = message.getDocument().getFullContentOctets();
             uid = message.getUid();
             this.exception = exception;
         }
